requirements.txt
pyarrow
apache-airflow-providers-apache-spark
pyspark==3.5.6